removed functions drawVertices/edges/vertexNumbers from main() to new header file
